Keep default cache builder in :persistent-cache
