+++ subprojects/core/src/main/java/org/gradle/cache/internal/DefaultUnscopedCacheBuilderFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,20 +15,10 @@
  */
 package org.gradle.cache.internal;
 
-import org.gradle.api.Action;
 import org.gradle.cache.CacheBuilder;
-import org.gradle.cache.CacheCleanupStrategy;
 import org.gradle.cache.UnscopedCacheBuilderFactory;
-import org.gradle.cache.FileLockManager;
-import org.gradle.cache.LockOptions;
-import org.gradle.cache.PersistentCache;
-import org.gradle.cache.internal.filelock.DefaultLockOptions;
 
 import java.io.File;
-import java.util.Collections;
-import java.util.Map;
-
-import static org.gradle.cache.internal.filelock.DefaultLockOptions.mode;
 
 public class DefaultUnscopedCacheBuilderFactory implements UnscopedCacheBuilderFactory {
     private final CacheFactory factory;
@@ -39,54 +29,7 @@ public class DefaultUnscopedCacheBuilderFactory implements UnscopedCacheBuilderF
 
     @Override
     public CacheBuilder cache(File baseDir) {
-        return new PersistentCacheBuilder(baseDir);
+        return new DefaultCacheBuilder(factory, baseDir);
     }
 
-    private class PersistentCacheBuilder implements CacheBuilder {
-        final File baseDir;
-        Map<String, ?> properties = Collections.emptyMap();
-        Action<? super PersistentCache> initializer;
-        CacheCleanupStrategy cacheCleanupStrategy;
-        LockOptions lockOptions = mode(FileLockManager.LockMode.Shared);
-        String displayName;
-
-        PersistentCacheBuilder(File baseDir) {
-            this.baseDir = baseDir;
-        }
-
-        @Override
-        public CacheBuilder withProperties(Map<String, ?> properties) {
-            this.properties = properties;
-            return this;
-        }
-
-        @Override
-        public CacheBuilder withDisplayName(String displayName) {
-            this.displayName = displayName;
-            return this;
-        }
-
-        @Override
-        public CacheBuilder withInitialLockMode(FileLockManager.LockMode mode) {
-            this.lockOptions = DefaultLockOptions.mode(mode);
-            return this;
-        }
-
-        @Override
-        public CacheBuilder withInitializer(Action<? super PersistentCache> initializer) {
-            this.initializer = initializer;
-            return this;
-        }
-
-        @Override
-        public CacheBuilder withCleanupStrategy(CacheCleanupStrategy cacheCleanupStrategy) {
-            this.cacheCleanupStrategy = cacheCleanupStrategy;
-            return this;
-        }
-
-        @Override
-        public PersistentCache open() {
-            return factory.open(baseDir, displayName, properties, lockOptions, initializer, cacheCleanupStrategy);
-        }
-    }
 }
